requirements_old.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
coloredlogs==15.0.1
Django==5.2
django-browser-reload==1.18.0
django-cors-headers==4.7.0
django-tailwind-4==0.1.4
djangorestframework==3.16.0
graph-retriever==0.8.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
langchain==0.3.23
langchain-chroma==0.2.3
langchain-community==0.3.21
langchain-core==0.3.52
langchain-graph-retriever==0.8.0
langchain-openai==0.3.13
langchain-text-splitters==0.3.8
langgraph==0.3.30
langgraph-checkpoint==2.0.24
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
SQLAlchemy==2.0.40
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
